+++ src/components/Modal/Modal.types.ts
@@ -24,7 +24,9 @@ export type Props = {
 	padding?: G.Responsive<number>;
 	active?: boolean;
 	transparentOverlay?: boolean;
+	autoFocus?: boolean;
 	onClose?: () => void;
+	ariaLabel?: string;
 	className?: G.ClassName;
 	overlayClassName?: G.ClassName;
 	attributes?: G.Attributes<"div", Props>;