+++ CHANGELOG.md
@@ -7,3 +7,5 @@
 - Badge: Relaxed the types to be able to pass dismissAriaLabel while onDismiss is undefined
 - Modal: Added full-screen position
 - Tooltip: Optional text prop
+- Modal: Added autoFocus=false support
+- Modal: Added ariaLabel support